+++ jaziku/core/settings/check.py
@@ -96,6 +96,27 @@ def configuration_run():
     env.config_run.settings['var_I_category_labels'] = format_labels(env.config_run.settings['var_I_category_labels'])
 
     # ------------------------
+    # calculation_mode_series_D
+    if env.config_run.settings['calculation_mode_series_D'] == 'default':
+        if env.var_D.TYPE_SERIES in env.var_D.CALCULATION_MODE_SERIES:
+            env.config_run.settings['calculation_mode_series_D'] = env.var_D.CALCULATION_MODE_SERIES[env.var_D.TYPE_SERIES]
+        else:
+            env.config_run.settings['calculation_mode_series_D'] = 'totalize'
+        env.globals_vars.input_settings["calculation_mode_series_D"] = env.config_run.settings['calculation_mode_series_D']
+    elif env.config_run.settings['calculation_mode_series_D'] in ['totalize', 'mean']:
+        if env.var_D.TYPE_SERIES in env.var_D.CALCULATION_MODE_SERIES and \
+            not env.var_D.CALCULATION_MODE_SERIES[env.var_D.TYPE_SERIES] == env.config_run.settings['calculation_mode_series_D']:
+            console.msg_error_configuration('calculation_mode_series_D',
+                                            _("For {0} the calculation_mode_series is '{1}', but\n"
+                                              "this should be '{2}' due the type of series.")
+                                            .format(env.var_D.TYPE_SERIES, env.config_run.settings['calculation_mode_series_D'],
+                                                    env.var_D.CALCULATION_MODE_SERIES[env.var_D.TYPE_SERIES]))
+        env.globals_vars.input_settings["calculation_mode_series_D"] = colored.green(env.config_run.settings['calculation_mode_series_D'])
+    else:
+        console.msg_error_configuration('calculation_mode_series_D', _("The calculation_mode_series_D is wrong, the options are:\n"
+                                                                       "default, totalize or mean"))
+
+    # ------------------------
     # limits var D
 
     # check and validate if file D is defined as particular file with
@@ -144,6 +165,27 @@ def configuration_run():
                    "a valid number, 'none' or 'default'.").format(env.config_run.settings['limits_var_D']['above'])))
 
     env.config_run.settings['limits_var_D']['ready'] = False
+
+    # ------------------------
+    # calculation_mode_series_I
+    if env.config_run.settings['calculation_mode_series_I'] == 'default':
+        if env.var_I.TYPE_SERIES in env.var_I.CALCULATION_MODE_SERIES:
+            env.config_run.settings['calculation_mode_series_I'] = env.var_I.CALCULATION_MODE_SERIES[env.var_I.TYPE_SERIES]
+        else:
+            env.config_run.settings['calculation_mode_series_I'] = 'totalize'
+        env.globals_vars.input_settings["calculation_mode_series_I"] = env.config_run.settings['calculation_mode_series_I']
+    elif env.config_run.settings['calculation_mode_series_I'] in ['totalize', 'mean']:
+        if env.var_I.TYPE_SERIES in env.var_I.CALCULATION_MODE_SERIES and \
+                not env.var_I.CALCULATION_MODE_SERIES[env.var_I.TYPE_SERIES] == env.config_run.settings['calculation_mode_series_I']:
+            console.msg_error_configuration('calculation_mode_series_I',
+                                            _("For {0} the calculation_mode_series is '{1}', but\n"
+                                              "this should be '{2}' due the type of series.")
+                                            .format(env.var_I.TYPE_SERIES, env.config_run.settings['calculation_mode_series_I'],
+                                                    env.var_I.CALCULATION_MODE_SERIES[env.var_I.TYPE_SERIES]))
+        env.globals_vars.input_settings["calculation_mode_series_I"] = colored.green(env.config_run.settings['calculation_mode_series_I'])
+    else:
+        console.msg_error_configuration('calculation_mode_series_I', _("The calculation_mode_series_I is wrong, the options are:\n"
+                                                                       "default, totalize or mean"))
 
     # ------------------------
     # limits var I