+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Changelog
 news
 ++++
 
+- Defined internal calculation_mode_series for var D and I in 'default' value and check it when is chose by the user
 - New options in runfile: calculation_mode_series_D and calculation_mode_series_I, these
   are the options to calculate the series: totalize or mean
 - Updated thresholds for var D for 3 and 7 categories (ATTENTION: this change modify several results)